+++ packages/messageport/CHANGELOG.md
@@ -1,3 +1,9 @@
 ## 0.1.0
 
 * Initial release.
+
+## 0.2.0
+
+* Deprecate `TizenMessagePort.createLocalPort` and `TizenMessagePort.connectToRemotePort`
+  in favor of newly added `LocalPort.create` and `RemotePort.connect`.
+* Minor cleanups.